split choose dialog, dynamic ui, eventbus, processengine
