+++ src/contracts/choose-dialog/IChooseDialogOption.ts
