+++ src/contracts/choose-dialog/index.ts
@@ -1,0 +1,1 @@
+export {IChooseDialogOption} from './IChooseDialogOption';
