+++ src/contracts/dynamic-ui/index.ts
@@ -1,0 +1,1 @@
+export {IDynamicUiService} from './IDynamicUiService';
