+++ src/contracts/eventbus/index.ts
@@ -1,0 +1,1 @@
+export {IEventBus} from './IEventBus';
