+++ src/contracts/index.ts
@@ -1,6 +1,6 @@
-export * from './iprocessengine';
+export * from './processengine/index';
 export * from './bpmnmodeler/index';
-export * from './idynamic-ui';
-export * from './ichoose-dialog';
+export * from './dynamic-ui/index';
+export * from './choose-dialog/index';
 export * from './authentication/index';
-export * from './ieventbus';
+export * from './eventbus/index';
