+++ src/contracts/processengine/IErrorResponse.ts
@@ -1,0 +1,3 @@
+export interface IErrorResponse {
+  error: any;
+}
