+++ src/contracts/processengine/IPagination.ts
@@ -1,0 +1,6 @@
+export interface IPagination<T> {
+  count: number;
+  offset: number;
+  limit: number;
+  data: Array<T>;
+}
