+++ src/contracts/processengine/index.ts
@@ -1,0 +1,5 @@
+export {IProcessDefEntity, IProcessEntity, IUserTaskEntity} from '@process-engine/process_engine_contracts';
+export {IErrorResponse} from './IErrorResponse';
+export {IPagination} from './IPagination';
+export {IProcessEngineRepository} from './IProcessEngineRepository';
+export {IProcessEngineService} from './IProcessEngineService';
